Flatten EventHandlerList
